update: Add ability for load balancer simulation to stop when request queue is empty and all requests have been handled
Add return values to servers when assigning requests and when attempting to resolve requests to understand the state of the server.

Have add feature to load balancer to keep track of the amount of available servers in its server group. Added a function to inquire if the whole server group is available.

Change load balancer simulation to add requests every 10-150 cycles instead of 10-40. Add functionality to exit simulation when request queue is empty and all requests have been resolved.

Update log file outputted by simulation.